rename types.ts to app-types.ts
