+++ grafana-plugin/src/app-types.ts
