+++ grafana-plugin/src/models/plugin/plugin.ts
@@ -1,8 +1,8 @@
+import { OnCallPluginMetaJSONData } from 'app-types';
 import { waitInMs } from 'helpers/async';
 import { AutoLoadingState } from 'helpers/decorators';
 import { isEqual } from 'lodash-es';
 import { makeAutoObservable, runInAction } from 'mobx';
-import { OnCallPluginMetaJSONData } from 'types';
 
 import { ActionKey } from 'models/loader/action-keys';
 import { GrafanaApiClient } from 'network/grafana-api/http-client';
