+++ grafana-plugin/src/module.ts
@@ -8,7 +8,7 @@ import { MobileAppConnectionWrapper } from 'containers/MobileAppConnection/Mobil
 import { PluginConfigPage } from 'containers/PluginConfigPage/PluginConfigPage';
 import { GrafanaPluginRootPage } from 'plugin/GrafanaPluginRootPage';
 
-import { OnCallPluginConfigPageProps, OnCallPluginMetaJSONData } from './types';
+import { OnCallPluginConfigPageProps, OnCallPluginMetaJSONData } from './app-types';
 
 const plugin = new AppPlugin<OnCallPluginMetaJSONData>().setRootPage(GrafanaPluginRootPage).addConfigPage({
   title: 'Configuration',
