+++ grafana-plugin/src/network/grafana-api/api.types.d.ts
@@ -1,4 +1,4 @@
-import { OnCallPluginMetaJSONData, OnCallPluginMetaSecureJSONData } from 'types';
+import { OnCallPluginMetaJSONData, OnCallPluginMetaSecureJSONData } from 'app-types';
 
 export type ServiceAccountDTO = {
   description: string;
